Modified NodeManager to only create its ZK node if it doesn't exist.
The #create_path method is called in quite a few places, often when the
node in question already exists. This causes a stream of NodeExists
exceptions from ZooKeeper, which might pollute its logs a bit.

I've modified the method to check if the node exists before creating it.
I didn't write specs, but manual testing showed this doesn't change redis_failover's
behavior